// packages/manager/apps/pci-savings-plan/src/components/CreatePlanForm/CreatePlan.type.ts
import { Technical } from '@/data/api/commercial-catalog';

export enum ResourceType {
  instance = 'instance',
  rancher = 'rancher',
}

export enum InstanceTechnicalName {
  b3 = 'b3',
  c3 = 'c3',
  r3 = 'r3',
  rancher = 'rancher',
}

export type InstanceInfo = {
  id: string;
  category: ResourceType;
  technicalName: InstanceTechnicalName;
  label: string;
  technical: {
    name: string;
    technical: Technical;
    hourlyPrice: number;
  }[];
};

export type Resource = {
  value: ResourceType;
  label: string;
  img: string;
};